mod02/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.1
plotly==5.20.0
matplotlib==3.8.4
scikit-learn==1.4.2
SQLAlchemy==2.0.29
